+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## []
 
+- Removed entity tooltip in REI
+- Fixed enchantment level coloring
 - Fixed double tooltip rendering in JEI
 
 ## [1.1.0]